configs/requirements.txt
# Configs 模块依赖
pydantic>=2.0.0
PyYAML>=6.0
omegaconf>=2.3.0
